+++ CHANGELOG.md
@@ -17,6 +17,8 @@ See [http://keepachangelog.com/](http://keepachangelog.com/) for best practices.
 ### Added
 - Multiple bug fixes for layout mistakes
 - Contributor permissiosn displayed in Contributor View page
+- Custom Author string in citation bug fix
+- Bibtx cutoff bug fix
 
 ## [v2.1.0] - 2015-08-04
 - Following functionality and lists.